requirements.txt
amqp==2.3.2
argon2-cffi==21.1.0
attrs==17.4.0
billiard==3.5.0.4
celery==4.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==6.7
dnspython==2.1.0
filelock==3.0.12
Flask==0.12.2
Flask-Login==0.4.1
flask-unsign==0.1.0
flask-unsign-wordlist==0.0.5
idna==2.10
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
meson==0.53.1
pluggy==0.6.0
py==1.5.2
py-altdns==1.0.0
pycparser==2.20
pyftpdlib==1.5.6
PyJWT==1.4.2
pytest==3.4.2
pytest-mock==1.7.1
pytz==2018.5
requests==2.25.1
requests-file==1.5.1
six==1.11.0
termcolor==1.1.0
tldextract==3.1.0
urllib3==1.26.5
vine==1.1.4
virtualenv==16.0.0
Werkzeug==0.14.1
